Move configure to tracing.py
